cpufreq_pwrutilx: update the governor
Based on schedutil, it uses a different formula to get next frequency.

In order for the governor to work well, there are some things which need to be set up accordingly. In /dev/cpuset, you need to set the cores file of foreground and top-app to 2-3, and cores file of background and system-bakcground to 0-1.

Trying very hard not to steer too far away from what schedutil tries to achieve.

Signed-off-by: ZeroInfinity <infinitezero345@gmail.com>
Change-Id: I8536aa6a561ea51d30c9f8852122050dea2b7ecf